==> 6_analyses_2020-2021.Rmd <==
---
title: "Response of Pinguicula vulgaris to geothermal heating"
subtitle: "Analyses 2020-2021"
author : "Alicia Valdés"
output:
  pdf_document:
    latex_engine: xelatex
    toc: yes
    toc_depth: 4
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
options(tibble.width = Inf)
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE) 
options("jtools-digits" = 3)
options(pillar.sigfigs=6)
```

```{r load packages, include=FALSE}
library(tidyverse)
library(readxl)
library(jtools)
library(RColorBrewer)
library(gridExtra)
library(ggeffects)
library(DHARMa)
library(ggthemes)
library(MASS)
library(glmmTMB)
library(broom.mixed)
library(parallel)
library(MuMIn)
library(aster)
library(knitr)
library(viridis)
library(cowplot)
```

```{r Define ggplot themes and palettes, include=FALSE}
my_theme <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(legend.position="none")+theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
my_theme_legend <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
myPalette <- colorRampPalette(brewer.pal(11, "YlOrRd"))
```

```{r load previously saved objects, include=FALSE}
load("output/BCIs_av_fitness_2020.RData")
load("output/BCIs_av_selection_2020.RData")
```

```{r include=FALSE}
ping_20_data <- read_csv("data/clean/ping_20_data.csv")
ping_21_data <- read_csv("data/clean/ping_21_data.csv")
```

```{r include=FALSE}
ping_20_data$ros_area <- with(ping_20_data,pi*dm1*dm2)
ping_21_data$ros_area <- with(ping_21_data,pi*dm1*dm2)
```

Define two variables: P (1 if pollinated, 0 if not) and F (1 if fed, 0 if not).

```{r}
ping_20_data <- ping_20_data %>%
  mutate(P=as.factor(ifelse(treatment=="P"|treatment=="PF",1,0)),
         F=as.factor(ifelse(treatment=="F"|treatment=="PF",1,0)))
```

```{r}
nrow(subset(ping_20_data,P==0))/nrow(ping_20_data)
nrow(subset(ping_20_data,F==0))/nrow(ping_20_data)
```

Approximately half of the plants are pollinated and half are fed.

Proportion of plants that produced 0 seeds:

```{r}
nrow(subset(ping_20_data,n_seeds==0))/nrow(ping_20_data)
```

Around half of the plants produced no seeds.

Create 0/1 variable for fitness (2020):

```{r}
ping_20_data$fitness_01<-with(ping_20_data,ifelse(n_seeds==0,0,
                                                  ifelse(n_seeds>0,1,NA)))
```

One case when n_seedpods>0 and n_seeds=0: convert to NA:

```{r}
subset(ping_20_data,n_seedpods>0&n_seeds==0)
```

```{r}
ping_20_data$n_seeds<-with(ping_20_data,ifelse(id=="P-450",NA,n_seeds))
```

Create 0/1 variable for fitness (2021):

```{r}
ping_21_data$fitness_01<-with(ping_21_data,ifelse(n_seedpods==0,0,
                                                  ifelse(n_seedpods>0,1,NA)))
```

# A. Analyses only 2020

## 1. Models for FFD 

Effect of temperature on FFD

Using linear models, including quadratic effects of temp.

```{r echo=TRUE}
FFD_2020_1<-lm(FFD_corr~temp+I(temp^2),ping_20_data)
summary(FFD_2020_1)
```

Model diagnostics (not shown) indicated that linear models are OK.

### Model prediction

```{r, fig.height=3, fig.width=4}
predict_FFD<-ggpredict(FFD_2020_1,terms="temp")
(fig1<-ggplot(ping_20_data,aes(x=temp,y=FFD_corr))+
  geom_ribbon(data=predict_FFD,
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="grey",alpha=0.7)+
  geom_line(data=predict_FFD,aes(x=x,y=predicted),size=1,color="black")+
  geom_point(size=2,alpha=0.2)+
  xlab("Soil temperature (ºC)")+ylab("FFD")+
  my_theme())
ggsave(filename="output/figures/fig1.tiff",device="tiff",
       width=9,height=8,units="cm",dpi=300,compression="lzw")
```

The effect looks almost linear.

Predictions of FFD\_corr for minimum and maximum temperatures:

```{r}
ggpredict(FFD_2020_1,terms="temp[minmax]") 
# 190.85-155.16=35.69 days earlier on warmer soils
```

## 2. Models for absolute fitness

### Model selection

```{r}
ping_20_data$ros_area_log<-log(ping_20_data$ros_area)
subset3<-subset(ping_20_data,!is.na(n_seeds)&!is.na(ros_area_log))
globmod_fitness_2020<-lm(n_seeds~temp*FFD_corr*P*F+ros_area_log,
                               subset3,na.action="na.fail")
clusterType <- if(length(find.package("snow", quiet = TRUE))) "SOCK" else "PSOCK"
clust1 <- try(makeCluster(getOption("cl.cores", 3), type = clusterType))
clusterExport(clust1, "subset3")
clusterEvalQ(clust1, library(stats))

modsel_fitness_2020<-pdredge(globmod_fitness_2020,fixed=c("ros_area_log"),
                             cluster=clust1)

summary(model.avg(modsel_fitness_2020,
                  subset=delta<2)) # Summary averaged model
importance(modsel_fitness_2020) # Variable importance
r.squaredGLMM(get.models(modsel_fitness_2020,subset=1)$"9102") #R square of best model
```

### Model diagnostics for best model

```{r}
plot(simulateResiduals(fittedModel=get.models(modsel_fitness_2020,
                                              subset=1)$"9102",n=5000))
```

### BCa intervals for averaged model

```{r eval=FALSE, include=FALSE}
av_fitness_2020<-lm(n_seeds~FFD_corr+P+temp+ros_area_log+
                          FFD_corr:P+FFD_corr:temp+P:temp+
                          FFD_corr:P:temp+F,
                        subset3,na.action="na.fail")

# FFD_corr
slp <- function(av_fitness_2020) coef(av_fitness_2020)[2]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# P1
slp <- function(av_fitness_2020) coef(av_fitness_2020)[3]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
P1_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# temp
slp <- function(av_fitness_2020) coef(av_fitness_2020)[4]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
temp_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# ros_area_log
slp <- function(av_fitness_2020) coef(av_fitness_2020)[5]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ros_area_log_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# FFD_corr:P1
slp <- function(av_fitness_2020) coef(av_fitness_2020)[7]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_P1_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# FFD_corr:temp
slp <- function(av_fitness_2020) coef(av_fitness_2020)[8]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_temp_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# P1:temp
slp <- function(av_fitness_2020) coef(av_fitness_2020)[9]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
P1_temp_ci <- as.data.frame(b1$bca[1,4:5])

# FFD_corr:P1:temp
slp <- function(av_fitness_2020) coef(av_fitness_2020)[10]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_P1_temp_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# F1
slp <- function(av_fitness_2020) coef(av_fitness_2020)[6]
b <- car::Boot(av_fitness_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
F1_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# Save confidence intervals as a table
BCIs_av_fitness_2020 <- cbind(rbind(FFD_corr_ci[1,],P1_ci[1,],
                                     temp_ci[1,],ros_area_log_ci[1,],
                                     FFD_corr_P1_ci[1,],
                                     FFD_corr_temp_ci[1,],
                                     P1_temp_ci[1,],FFD_corr_P1_temp_ci[1,],
                                     F1_ci[1,]),
                               rbind(FFD_corr_ci[2,],P1_ci[2,],
                                     temp_ci[2,],ros_area_log_ci[2,],
                                     FFD_corr_P1_ci[2,],
                                     FFD_corr_temp_ci[2,],
                                     P1_temp_ci[2,],FFD_corr_P1_temp_ci[2,],
                                     F1_ci[2,]))
colnames(BCIs_av_fitness_2020)<-c("lower","upper")
rownames(BCIs_av_fitness_2020) <- c("FFD_corr","P1","temp",
                                     "ros_area_log","FFD_corr:P1",
                                     "FFD_corr:temp","P1:temp",
                                     "FFD_corr:P1:temp","F1")
save(BCIs_av_fitness_2020,file="output/BCIs_av_fitness_2020.RData")
```

```{r}
BCIs_av_fitness_2020
```

### Model predictions from best model

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
               terms=c("FFD_corr[all]","temp")))
```

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
               terms=c("temp","P")))
```

3-way interaction (two different representations):

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
               terms=c("FFD_corr","temp","P")))
```

For non-pollinated plants, it is always better to flower early, irrespective of soil temperatures. For pollinated plants, it is better to flower early at cold soil temperatures, but late at medium and warm soil temperatures. 

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_fitness_2020,
                          subset=1)$"9102",terms=c("FFD_corr","P","temp")))
```

At low temperatures (left panel above), it is always better to flower early, irrespective of if the plant was pollinated or not. At medium and high temperatures (center and right panel above), it is better to flower early for non-pollinated plants, but better to flower late for pollinated plants.

```{r, fig.height=3, fig.width=6}
predict_fitness1<-ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
               terms=c("FFD_corr[all]","temp[all]"))
ggplot(predict_fitness1,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  xlab("FFD")+ylab("Predicted fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=132,begin=0.4,end=0.9,direction=-1))+
  labs(colour="Soil  temperature (ºC)")
ggsave(filename="output/figures/fitness_1.tiff",device="tiff",
       width=15,height=8,units="cm",dpi=300,compression="lzw")

predict_fitness2<-ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
               terms=c("temp[all]","FFD_corr[all]"))
ggplot(predict_fitness2,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  xlab("Soil temperature (ºC)")+ylab("Predicted fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=132,begin=0.4,end=0.9,direction=-1))+
  labs(colour="FFD")
ggsave(filename="output/figures/fitness_2.tiff",device="tiff",
       width=11,height=8,units="cm",dpi=300,compression="lzw")
```

```{r, fig.height=3, fig.width=6}
predict_fitness3<-ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
               terms=c("temp[all]","P"))
ggplot(predict_fitness3,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_ribbon(aes(fill=group),alpha=0.2)+
  geom_line(aes(color=group),size=1)+
  xlab("FFD")+ylab("Predicted fitness")+
  scale_fill_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  scale_color_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  my_theme_legend()
ggsave(filename="output/figures/fitness_3.tiff",device="tiff",
       width=14,height=8,units="cm",dpi=300,compression="lzw")
```

```{r}
predict_fitness4<-ggpredict(get.models(modsel_fitness_2020,subset=1)$"9102",
                            terms=c("FFD_corr[all]","temp[all]","P"))
P.labs <- c("Non-pollinated","Pollinated")
names(P.labs) <- c(0,1)
ggplot(predict_fitness4,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  facet_wrap(~facet,labeller=labeller(facet=P.labs))+
  xlab("FFD")+ylab("Predicted fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=132,begin=0.4,end=0.9,direction=-1))+
  labs(colour="Soil  temperature (ºC)")
ggsave(filename="output/figures/fitness_4.tiff",device="tiff",
       width=20,height=8,units="cm",dpi=300,compression="lzw")

predict_fitness5<-ggpredict(get.models(modsel_fitness_2020,
                          subset=1)$"9102",terms=c("FFD_corr[all]","P","temp"))
temp.labs <- c("9.95 ºC","13.46 ºC","16.96 ºC")
names(temp.labs) <- c(9.95,13.46,16.96)
ggplot(predict_fitness5,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=group),size=1)+
  geom_ribbon(aes(fill=group),alpha=0.2)+
  facet_wrap(~facet,labeller=labeller(facet=temp.labs))+
  xlab("FFD")+ylab("Predicted fitness")+
  scale_fill_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  scale_color_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  my_theme_legend()+
  labs(colour=NULL,fill=NULL)
ggsave(filename="output/figures/fitness_5.tiff",device="tiff",
       width=20,height=8,units="cm",dpi=300,compression="lzw")
```

## 3. Models for relative fitness (selection models)

Standardize traits and relativize fitness:

```{r}
# Standardize FFD
ping_20_data$FFD_corr_std<-as.numeric(scale(ping_20_data$FFD_corr))
# Standardize rosette area (first log, then scale)
ping_20_data$ros_area_std<-as.numeric(scale(log(ping_20_data$ros_area)))
# Relativize fitness
ping_20_data$n_seeds_rel<-with(ping_20_data,n_seeds/mean(n_seeds,na.rm=T))
```

### Model selection

```{r}
subset2<-subset(ping_20_data,!is.na(n_seeds_rel)&!is.na(ros_area_std))
globmod_selection_2020<-lm(n_seeds_rel~temp*FFD_corr_std*P*F+ros_area_std,
                               subset2,na.action="na.fail")
clusterType <- if(length(find.package("snow", quiet = TRUE))) "SOCK" else "PSOCK"
clust1 <- try(makeCluster(getOption("cl.cores", 3), type = clusterType))
clusterExport(clust1, "subset2")
clusterEvalQ(clust1, library(stats))

modsel_selection_2020<-pdredge(globmod_selection_2020,fixed=c("ros_area_std"),
                                 cluster=clust1)

summary(model.avg(modsel_selection_2020,
                  subset=delta<2)) # Summary averaged model
importance(modsel_selection_2020) # Variable importance
r.squaredGLMM(get.models(modsel_selection_2020,subset=1)$"9102") #R square of best model
```

### Model diagnostics for best model

```{r}
plot(simulateResiduals(fittedModel=get.models(modsel_selection_2020,
                                              subset=1)$"9102",n=5000))
```
### BCa intervals for averaged model

```{r eval=FALSE, include=FALSE}
av_selection_2020<-lm(n_seeds_rel~FFD_corr_std+P+temp+ros_area_std+
                          FFD_corr_std:P+FFD_corr_std:temp+P:temp+
                          FFD_corr_std:P:temp+F,
                        subset2,na.action="na.fail")

# FFD_corr_std
slp <- function(av_selection_2020) coef(av_selection_2020)[2]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# P1
slp <- function(av_selection_2020) coef(av_selection_2020)[3]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
P1_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# temp
slp <- function(av_selection_2020) coef(av_selection_2020)[4]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
temp_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# ros_area_std 
slp <- function(av_selection_2020) coef(av_selection_2020)[5]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
ros_area_std_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# FFD_corr_std:P1
slp <- function(av_selection_2020) coef(av_selection_2020)[7]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_std_P1_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# FFD_corr_std:temp
slp <- function(av_selection_2020) coef(av_selection_2020)[8]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_std_temp_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# P1:temp
slp <- function(av_selection_2020) coef(av_selection_2020)[9]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
P1_temp_ci <- as.data.frame(b1$bca[1,4:5])

# FFD_corr_std:P1:temp
slp <- function(av_selection_2020) coef(av_selection_2020)[10]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
FFD_corr_std_P1_temp_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# F1
slp <- function(av_selection_2020) coef(av_selection_2020)[6]
b <- car::Boot(av_selection_2020,slp, R=10000) # note the capital B
b1 <- boot::boot.ci(b,type="bca")
F1_ci <- as.data.frame(b1$bca[1,4:5])
rm(slp, b, b1)

# Save confidence intervals as a table
BCIs_av_selection_2020 <- cbind(rbind(FFD_corr_std_ci[1,],P1_ci[1,],
                                     temp_ci[1,],ros_area_std_ci[1,],
                                     FFD_corr_std_P1_ci[1,],
                                     FFD_corr_std_temp_ci[1,],
                                     P1_temp_ci[1,],FFD_corr_std_P1_temp_ci[1,],
                                     F1_ci[1,]),
                               rbind(FFD_corr_std_ci[2,],P1_ci[2,],
                                     temp_ci[2,],ros_area_std_ci[2,],
                                     FFD_corr_std_P1_ci[2,],
                                     FFD_corr_std_temp_ci[2,],
                                     P1_temp_ci[2,],FFD_corr_std_P1_temp_ci[2,],
                                     F1_ci[2,]))
colnames(BCIs_av_selection_2020)<-c("lower","upper")
rownames(BCIs_av_selection_2020) <- c("FFD_corr_std","P1","temp",
                                     "ros_area","FFD_corr_std:P1",
                                     "FFD_corr_std:temp","P1:temp",
                                     "FFD_corr:P1:temp","F1")
save(BCIs_av_selection_2020,file="output/BCIs_av_selection_2020.RData")
```

```{r}
BCIs_av_selection_2020
```

### Model predictions from best model

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("FFD_corr_std[all]","temp")))
```

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("temp","P")))
```
3-way interaction (two different representations):

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("FFD_corr_std","temp","P")))
```

For non-pollinated plants, it is always better to flower early, irrespective of soil temperatures. For pollinated plants, it is better to flower early at cold soil temperatures, but late at medium and warm soil temperatures. 

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_selection_2020,
                          subset=1)$"9102",terms=c("FFD_corr_std","P","temp")))
```

At low temperatures (left panel above), it is always better to flower early, irrespective of if the plant was pollinated or not. At medium and high temperatures (center and right panel above), it is better to flower early for non-pollinated plants, but better to flower late for pollinated plants.

```{r, fig.height=3, fig.width=6}
predict_relfitness1<-ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("FFD_corr_std[all]","temp[all]"))
ggplot(predict_relfitness1,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  xlab("FFD (standardized)")+ylab("Predicted relative fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=132,begin=0.4,end=0.9,direction=-1))+
  labs(colour="Soil  temperature (ºC)")
ggsave(filename="output/figures/relfitness_1.tiff",device="tiff",
       width=15,height=8,units="cm",dpi=300,compression="lzw")

predict_relfitness2<-ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("temp[all]","FFD_corr_std[all]"))
ggplot(predict_relfitness1,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  xlab("Soil temperature (ºC)")+ylab("Predicted relative fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=132,begin=0.4,end=0.9,direction=-1))+
  labs(colour="FFD\n(standardized)")
ggsave(filename="output/figures/relfitness_2.tiff",device="tiff",
       width=13,height=8,units="cm",dpi=300,compression="lzw")
```

```{r, fig.height=3, fig.width=6}
predict_relfitness3<-ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("temp[all]","P"))
ggplot(predict_relfitness3,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_ribbon(aes(fill=group),alpha=0.2)+
  geom_line(aes(color=group),size=1)+
  xlab("Soil temperature (ºC)")+ylab("Predicted relative fitness")+
  scale_fill_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  scale_color_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  my_theme()
ggsave(filename="output/figures/relfitness_3.tiff",device="tiff",
       width=10,height=8,units="cm",dpi=300,compression="lzw")
```

```{r}
predict_relfitness4<-ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
                            terms=c("FFD_corr_std[all]","temp[all]","P"))
ggplot(predict_relfitness4,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  facet_wrap(~facet,labeller=labeller(facet=P.labs))+
  xlab("FFD (standardized)")+ylab("Predicted relative fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=132,begin=0.4,end=0.9,direction=-1))+
  labs(colour="Soil  temperature (ºC)")
ggsave(filename="output/figures/relfitness_4.tiff",device="tiff",
       width=20,height=8,units="cm",dpi=300,compression="lzw")

predict_relfitness5<-ggpredict(get.models(modsel_selection_2020,
                          subset=1)$"9102",terms=c("FFD_corr_std[all]","P","temp"))
ggplot(predict_relfitness5,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=group),size=1)+
  geom_ribbon(aes(fill=group),alpha=0.2)+
  facet_wrap(~facet,labeller=labeller(facet=temp.labs))+
  xlab("FFD (standardized)")+ylab("Predicted relative fitness")+
  scale_fill_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  scale_color_discrete(name=NULL,labels=c("Non-pollinated","Pollinated"))+
  my_theme_legend()+
  labs(colour=NULL,fill=NULL)
ggsave(filename="output/figures/relfitness_5.tiff",device="tiff",
       width=20,height=8,units="cm",dpi=300,compression="lzw")
```

Predictions for minimum and maximum temperatures:

```{r}
ggpredict(get.models(modsel_selection_2020,subset=1)$"9102",
               terms=c("temp[minmax]","P"))
# 190.85-155.16=35.69 days earlier on warmer soils
```

## 4. Models for binomial fitness (maybe not included in paper)

### Model selection

```{r}
subset1<-subset(ping_20_data,!is.na(fitness_01)&!is.na(ros_area_std))
globmod_selection_2020_01<-glm(fitness_01~temp*FFD_corr_std*P*F+ros_area_std,
                               subset1,family="binomial",na.action="na.fail")
clusterType <- if(length(find.package("snow", quiet = TRUE))) "SOCK" else "PSOCK"
clust1 <- try(makeCluster(getOption("cl.cores", 3), type = clusterType))
clusterExport(clust1, "subset1")
clusterEvalQ(clust1, library(stats))

modsel_selection_2020_01<-pdredge(globmod_selection_2020_01,
                                 fixed=c("ros_area_std"),
                                 cluster=clust1)

summary(model.avg(modsel_selection_2020_01,
                  subset=delta<2)) # Summary averaged model
importance(modsel_selection_2020_01) # Variable importance
r.squaredGLMM(get.models(modsel_selection_2020_01,subset=1)$"266") #R square of best model
```

### Model diagnostics for best model

```{r}
plot(simulateResiduals(fittedModel=get.models(modsel_selection_2020_01,
                                              subset=1)$"266",n=5000))
```

### Model predictions from best model

```{r,fig.height=3, fig.width=4}
plot(ggpredict(get.models(modsel_selection_2020_01,subset=1)$"266",
               terms=c("FFD_corr_std[all]","temp")))
```

We could use the method described by Janzen and Stern (1998) to transform the logit coefficients to linear scale, but not sure that is needed. 
Some code for that here: https://github.com/mabarbour/complexity_selection/blob/master/analyses/reproduce_analyses.Rmd

## 5. Aster models

Any Fruits (seedpods_01) --> Fruits (n_seedpods) --> Seeds (n_seeds)

Any fruits (did the plant produce fruits) is Bernoulli
Number of Fruits is zero-truncated negative binomial
Number of seeds is Poisson 

Reshape data:

```{r}
ping_20_data$n_seedpods<-as.integer(ping_20_data$n_seedpods)
ping_20_data$n_seeds<-as.integer(ping_20_data$n_seeds)
ping_20_data$seedpods_01<-with(ping_20_data,ifelse(n_seedpods>0,1,0))
vars <- c("seedpods_01", "n_seedpods","n_seeds") # Our fitness variables
redata <- reshape(data.frame(subset(ping_20_data,
                                    !is.na(n_seeds)&!is.na(ros_area_std))
                             [c(1:2,10,15:16,21:29)]), 
                  varying = list(vars),
                  direction = "long", 
                  timevar = "varb", # reshape to long format
                  times = as.factor(vars), v.names = "resp")
redata <- data.frame(redata, root = 1) # This adds a variable root to the df
# and makes all its values one (including for non-root nodes, but those values
# are ignored by all aster package functions).

# resp contains all of the data in the variables indicated by the string
# vars packed into a single vector
# varb indicates which original variable the corresponding element of
# resp came from
```

```{r}
names(data.frame(subset(ping_20_data,
                                    !is.na(n_seeds)&!is.na(ros_area_std))
                             [c(1:2,10,15:16,21:29)]))
names(redata)
nrow(data.frame(subset(ping_20_data,
                                    !is.na(n_seeds)&!is.na(ros_area_std))
                             [c(1:2,10,15:16,21:29)]))
nrow(redata)
nrow(data.frame(subset(ping_20_data,
                                    !is.na(n_seeds)&!is.na(ros_area_std))
                             [c(1:2,10,15:16,21:29)])) * length(vars)
sapply(redata, class)
levels(redata$varb)
length(unique(redata$id))
```

Show the graphical model:

```{r}
pred <- c(0,1,2) # specifies the predecessor structure of the graph

foo <- c("root", vars)
pvars <- foo[pred + 1]
bar <- cbind(pvars, vars)
colnames(bar) <- c("pred", "succ")
bar
```
Look at Poisson distributions for seedpods and seeds:

```{r}
#For n_seedpods
n_seedpods.dist <- subset(ping_20_data,
                     !is.na(n_seeds)&
                       !is.na(ros_area_std))$n_seedpods[subset(ping_20_data,
        !is.na(n_seeds)&!is.na(ros_area_std))$seedpods_01 == 1]
hist(n_seedpods.dist)
length(n_seedpods.dist) #130
sum(n_seedpods.dist == 0) #0
n_seedpods.parms <- fitdistr(n_seedpods.dist, "poisson")
n_seedpods.parms #lambda = 1.14615385

#For n_seeds
n_seeds.dist <- subset(ping_20_data,
                     !is.na(n_seeds)&
                       !is.na(ros_area_std))$n_seeds[subset(ping_20_data,
                     !is.na(n_seeds)&
                       !is.na(ros_area_std))$n_seedpods > 0]
hist(n_seeds.dist) 
length(n_seeds.dist) #130
sum(n_seeds.dist == 0) #0
n_seeds.parms <-fitdistr(n_seeds.dist, "poisson")
n_seeds.parms # lambda = 234.069231
```

Look at negative binomial distributions for seedpods and seeds:

```{r}
#For n_seedpods
n_seedpods.parms2 <- fitdistr(n_seedpods.dist, "negative binomial")
n_seedpods.parms2 #size = 4.876975e+02

#For n_seeds
n_seeds.parms2 <-fitdistr(n_seeds.dist, "negative binomial")
n_seeds.parms2 # size = 5.8823340
```
Specifying families:

```{r}
fam <- c(1,2,3) # specifies the one-parameter exponential families for the nodes
famlist <- list(fam.bernoulli(), 
                fam.truncated.negative.binomial(size=4.876975e+02,
                                                 truncation = 0),
                fam.poisson())
# Here: Bernouilli, truncated.negative.binomial, poisson

vars[fam == 1]
vars[fam == 2]
vars[fam == 3]
```

```{r}
# creating the fitness variable of interest
seeds <- grep("n_seeds", as.character(redata$varb))
seeds <- is.element(seq(along = redata$varb), seeds)
redata <- data.frame(redata, seeds = as.integer(seeds))
names(redata)
```

Center the predictors (I think we do not need to use standardized predictors in Aster models, but centering the predictors helps the model to converge):

```{r}
redata$temp_c<-redata$temp-mean(redata$temp)
redata$FFD_c<-redata$FFD_corr-mean(redata$FFD_corr)
redata$ros_area_c<-redata$ros_area-mean(redata$ros_area)
```

Fitting the full Aster model:

```{r}
# Full Aster model with seeds as the interaction for all fixed variables
aster1 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
# Include varb in the formula to get a different “intercept” 
# for each node in the graph (i.e. one for fitness_01 and one for n_seeds)
# The categorical variable varb gets turned into as many dummy
# variables as there are nodes in the graph, one is dropped, and the
# “intercept” dummy variable (all components = 1) is added; the
# effect is to provide a different intercept for each node.
# Include interaction seeds: with all other variables
# Seehttps://groups.google.com/g/aster-analysis-user-group/c/xwanCX8VOWQ/m/AcTnF3E3BAAJ 
# and code there.
# According to https://www.stat.umn.edu/geyer/8931aster/slides/s1.pdf#page=55 
# (slides 55-71), we do this to have separate coefficients for the “layer” 
# of the graph consisting of terminal nodes (as in aster models, regression
# coefficients “for” a node of the graph also influence all “earlier” nodes
# To estimate the effects of each predictor on lifetime fitness, each predictor
# was fit at the level of total seed set in the model (Shaw et al. 2008)
```

```{r}
summary(aster1, 
        show.graph = TRUE, # Table about the graph structure in the printout
        info.tol = 1e-11) # To help the model to converge
# (see https://groups.google.com/g/aster-analysis-user-group/c/xXXROez8aAA/m/BrNE6whiIEQJ)
```

### Test interactions

The Aster manual advises to not look at significances in the summary, but instead compare models with a likelihood ratio test to evaluate significance of each term. The comparison between two or more models . . . will only be valid if they are (1) fitted to the same dataset, (2) models are nested, (3) models are of the same type (all conditional or all unconditional), (4) have the same dependence graph and exponential families. None of this is currently checked.

Determine the significance of fixed effects and interactions by comparing submodels without the term of interest to the full model using LRTs.

Use LRTs comparing submodels to fuller models to test each predictor of interest.

Full model with different syntax, making it easier to remove terms:

```{r}
aster1 <- aster(resp ~ varb+seeds:temp_c+seeds:FFD_c+seeds:P+seeds:F+
                      seeds:ros_area_c+seeds:temp_c:FFD_c+seeds:temp_c:P+
                      seeds:FFD_c:P+seeds:temp_c:F+seeds:FFD_c:F+seeds:P:F+
                      seeds:temp_c:FFD_c:P+seeds:temp_c:FFD_c:F+
                      seeds:temp_c:P:F+seeds:FFD_c:P:F+seeds:temp_c:FFD_c:P:F,
                pred, fam, famlist=famlist, varb, id, root, data = redata)
```

Submodels:

Testing the significance of seeds:temp_c:FFD_c:P:F

```{r}
aster1_1 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                        seeds:temp_c:FFD_c:P:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_1,aster1)
```

Testing the significance of seeds:FFD_c:P:F

```{r}
aster1_2 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_2,aster1_1)
```

Testing the significance of seeds:temp_c:P:F

```{r}
aster1_3 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_3,aster1_2)
```

Testing the significance of seeds:temp_c:FFD_c:F

```{r}
aster1_4 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_4,aster1_3)
```

Testing the significance of seeds:temp_c:FFD_c:P

```{r}
aster1_5 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_5,aster1_4)
```

Testing the significance of seeds:P:F

```{r}
aster1_6 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P-
                    seeds:P:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_6,aster1_5)
```

Testing the significance of seeds:FFD_c:F

```{r}
aster1_7 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P-
                    seeds:P:F-
                    seeds:FFD_c:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_7,aster1_6)
```

Testing the significance of seeds:temp_c:F

```{r}
aster1_8 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P-
                    seeds:P:F-
                    seeds:FFD_c:F-
                    seeds:temp_c:F,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_8,aster1_7)
```

Testing the significance of seeds:FFD_c:P

```{r}
aster1_9 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P-
                    seeds:P:F-
                    seeds:FFD_c:F-
                    seeds:temp_c:F-
                    seeds:FFD_c:P,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_9,aster1_8)
```

Testing the significance of seeds:temp_c:P

```{r}
aster1_10 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P-
                    seeds:P:F-
                    seeds:FFD_c:F-
                    seeds:temp_c:F-
                    seeds:FFD_c:P-
                    seeds:temp_c:P,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_10,aster1_9)
```

Testing the significance of seeds:temp_c:FFD_c

```{r}
aster1_11 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P*F+ros_area_c)-
                    seeds:temp_c:FFD_c:P:F-
                    seeds:FFD_c:P:F-
                    seeds:temp_c:P:F-
                    seeds:temp_c:FFD_c:F-
                    seeds:temp_c:FFD_c:P-
                    seeds:P:F-
                    seeds:FFD_c:F-
                    seeds:temp_c:F-
                    seeds:FFD_c:P-
                    seeds:temp_c:P-
                    seeds:temp_c:FFD_c,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_11,aster1_10)
```

### Test main effects without interactions

```{r}
aster1_mains <- aster(resp ~ varb+seeds:(temp_c+FFD_c+P+F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
summary(aster1_mains, 
        show.graph = TRUE, # Table about the graph structure in the printout
        info.tol = 1e-11) # To help the model to converge
```

Submodels:

Testing the significance of seeds:ros_area_c

```{r}
aster1_12 <- aster(resp ~ varb+seeds:(temp_c+FFD_c+P+F),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_12,aster1_mains)
```

Testing the significance of seeds:F

```{r}
aster1_13 <- aster(resp ~ varb+seeds:(temp_c+FFD_c+P+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_13,aster1_mains)
```

Testing the significance of seeds:P

```{r}
aster1_14 <- aster(resp ~ varb+seeds:(temp_c+FFD_c+F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_14,aster1_mains)
```

Testing the significance of seeds:FFD_c

```{r}
aster1_15 <- aster(resp ~ varb+seeds:(temp_c+P+F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_15,aster1_mains)
```

Testing the significance of seeds:temp_c

```{r}
aster1_16 <- aster(resp ~ varb+seeds:(FFD_c+P+F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster1_16,aster1_mains)
```

### Table LRTs

```{r message=FALSE, warning=FALSE}
table_LRTs<-bind_rows(
  tidy(anova(aster1_16,aster1_mains))%>%mutate(term="seeds:temp_c"),
  tidy(anova(aster1_15,aster1_mains))%>% mutate(term="seeds:FFD_c"),
  tidy(anova(aster1_14,aster1_mains))%>%mutate(term="seeds:P"),
  tidy(anova(aster1_13,aster1_mains))%>%mutate(term="seeds:F"),
  tidy(anova(aster1_12,aster1_mains))%>%mutate(term="seeds:ros_area_c"),
  tidy(anova(aster1_11,aster1_10))%>%mutate(term="seeds:temp_c:FFD_c"),
  tidy(anova(aster1_10,aster1_9))%>% mutate(term="seeds:temp_c:P"),
  tidy(anova(aster1_9,aster1_8))%>%mutate(term="seeds:FFD_c:P"),
  tidy(anova(aster1_8,aster1_7))%>%mutate(term="seeds:temp_c:F"),
  tidy(anova(aster1_7,aster1_6))%>%mutate(term="seeds:FFD_c:F"),
  tidy(anova(aster1_6,aster1_5))%>%mutate(term="seeds:P:F"),
  tidy(anova(aster1_5,aster1_4))%>%mutate(term="seeds:temp_c:FFD_c:P"),
  tidy(anova(aster1_4,aster1_3))%>%mutate(term="seeds:temp_c:FFD_c:F"),
  tidy(anova(aster1_3,aster1_2))%>%mutate(term="seeds:temp_c:P:F"),
  tidy(anova(aster1_2,aster1_1))%>%mutate(term="seeds:FFD_c:P:F"),
  tidy(anova(aster1_1,aster1))%>%mutate(term="seeds:temp_c:FFD_c:P:F")
)
kable(table_LRTs)
```

Significant effects are seeds:temp_c:FFD_c and seeds:temp_c:P.
Interaction seeds:FFD_c:P:temp_c is not significant (different from tresults of selection models).

### Selection gradients

If we need to calculate selection coefficients (gradients), we can use code in Peschel et al. 2021, and in https://groups.google.com/g/aster-analysis-user-group/c/6UlcfGBWM4s/m/eseUuoG3hmYJ

```{r}
# Predict the expected fitness for each individual from the aster model
paster1 <- predict(aster1)
paster1 <- matrix(paster1, nrow = nrow(aster1$x), ncol = ncol(aster1$x))
colnames(paster1) <- colnames(aster1$x)
mufit <- paster1[, grep("n_seeds", colnames(paster1))] # Expected fitness

# Calculate relative fitness (fitness~i~ / mean fitness) for each individual
wmu <- mufit/mean(mufit)

# Fit OLS regression with traits as predictors. 
# Coefficients are the selection gradients.
wmout <- lm(wmu ~ temp_c*FFD_c*P*F+ros_area_c,
            bind_cols(subset(ping_20_data,!is.na(n_seeds)&!is.na(ros_area_std)),
                      wmu=wmu)%>%
              mutate(temp_c=temp-mean(temp),
                     FFD_c=FFD_corr-mean(FFD_corr),
                     ros_area_c=ros_area-mean(ros_area)))
summary(wmout)
```

Calculating the *correct* standard errors is a bit more complicated but can be done following the code starting on p5 of TR #675( http://conservancy.umn.edu/bitstream/11299/56394/1/betaTR.pdf).

I am not sure we need to include this in the paper.

### Reduced model without F

```{r}
aster2 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
summary(aster2,show.graph = TRUE,info.tol = 1e-11) 
```

#### Test some interactions

Submodels:

Testing the significance of seeds:temp_c:FFD_c:P

```{r}
aster2_1 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P+ros_area_c)-
                        seeds:temp_c:FFD_c:P,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster2_1,aster2)
```

Testing the significance of seeds:FFD_c:P

```{r}
aster2_2 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P+ros_area_c)-
                    seeds:temp_c:FFD_c:P-
                    seeds:FFD_c:P,
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster2_2,aster2_1)
```

Significances similar to full model, provide results with reduced model in Appendix (with full table LRTs).

### Reduced model with terms included in averaged LMs (USE in main text)

```{r}
aster3 <- aster(resp ~ varb+seeds:(temp_c*FFD_c*P+F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
summary(aster3,show.graph = TRUE,info.tol = 1e-11) 
```

#### Test interactions

Full model with different syntax, making it easier to remove terms:

```{r}
aster3 <- aster(resp ~ varb+seeds:temp_c+seeds:FFD_c+seeds:P+seeds:F+
                      seeds:ros_area_c+seeds:temp_c:FFD_c+seeds:temp_c:P+
                      seeds:FFD_c:P+seeds:temp_c:FFD_c:P,
                pred, fam, famlist=famlist, varb, id, root, data = redata)
```

Submodels:

Testing the significance of seeds:temp_c:FFD_c:P

```{r}
aster3_1 <- aster(resp ~ varb+seeds:temp_c+seeds:FFD_c+seeds:P+seeds:F+
                      seeds:ros_area_c+seeds:temp_c:FFD_c+seeds:temp_c:P+
                      seeds:FFD_c:P,
                pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_1,aster3)
```

Testing the significance of seeds:FFD_c:P

```{r}
aster3_2 <- aster(resp ~ varb+seeds:temp_c+seeds:FFD_c+seeds:P+seeds:F+
                      seeds:ros_area_c+seeds:temp_c:FFD_c+seeds:temp_c:P,
                pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_2,aster3_1)
```

Testing the significance of seeds:temp_c:P

```{r}
aster3_3 <- aster(resp ~ varb+seeds:temp_c+seeds:FFD_c+seeds:P+seeds:F+
                      seeds:ros_area_c+seeds:temp_c:FFD_c,
                pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_3,aster3_2)
```

Testing the significance of seeds:temp_c:FFD_c

```{r}
aster3_4 <- aster(resp ~ varb+seeds:temp_c+seeds:FFD_c+seeds:P+seeds:F+
                      seeds:ros_area_c,
                pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_4,aster3_3)
```

#### Test main effects without interactions

```{r}
aster3_mains <- aster(resp ~ varb+seeds:(temp_c+FFD_c+P+F+ros_area_c),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
summary(aster3_mains, 
        show.graph = TRUE, # Table about the graph structure in the printout
        info.tol = 1e-11) # To help the model to converge
```

Submodels:

Testing the significance of seeds:ros_area_c

```{r}
aster3_5 <- aster(resp ~ varb+seeds:(temp_c+FFD_c+P+F),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_5,aster3_mains)
```

Testing the significance of seeds:F

```{r}
aster3_6 <- aster(resp ~ varb+seeds:(temp_c+FFD_c+P+ros_area),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_6,aster3_mains)
```

Testing the significance of seeds:P

```{r}
aster3_7 <- aster(resp ~ varb+seeds:(temp_c+FFD_c+F+ros_area),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_7,aster3_mains)
```

Testing the significance of seeds:FFD_c

```{r}
aster3_8 <- aster(resp ~ varb+seeds:(temp_c+P+F+ros_area),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_8,aster3_mains)
```

Testing the significance of seeds:temp_c

```{r}
aster3_9 <- aster(resp ~ varb+seeds:(FFD_c+P+F+ros_area),
                    pred, fam, famlist=famlist, varb, id, root, data = redata)
```

```{r}
anova(aster3_9,aster3_mains)
```

### Predictions

#### seeds:temp_c:P

```{r}
# make prediction df
aster_predict1 <- data.frame(
  expand_grid(temp_c = seq(from = -5, to = 13, length.out = 100),
              P = as.factor(c(0,1))),
  FFD_c = 0, # centred so mean=0
  ros_area_c = 0, # centred so mean=0
  #P = as.factor(0), # no pollination
  F = as.factor(0),
  seedpods_01 = 1,
  n_seedpods = 1,
  n_seeds = 0,
  root = 1
)

# reshape data to long format
aster_predict1_long <- reshape(as.data.frame(aster_predict1), 
                              varying = list(vars), direction = "long", 
                              timevar = "varb", times = as.factor(vars), 
                              v.names = "resp")

# add artifice
aster_predict1_long$seeds <- as.integer(ifelse(
  aster_predict1_long$varb == "n_seeds", 1, 0))

aster_predict1.df <- aster_predict1_long

# fixed model predictions
aster3.p1 <- predict(aster3, newdata = aster_predict1_long,
                    varvar = varb,idvar = id, root = root,
                    info.tol = 1e-11, se.fit = T)

aster_predict1.df$fit <- aster3.p1$fit
aster_predict1.df$se <- aster3.p1$se.fit

aster_predict1.df.n_seed <- aster_predict1.df %>%
  filter(varb == "n_seeds")

aster_predict1.df.n_seed<-aster_predict1.df.n_seed%>%
  mutate(ymin=fit-se,ymax=fit+se)

P.labs <- c("Non-supplemented","Supplemented")
names(P.labs) <- c(0,1)
ggplot(aster_predict1.df.n_seed)+
  geom_ribbon(aes(x=temp_c,ymin=ymin,ymax=ymax,fill=P),alpha=0.25)+
  geom_line(aes(x=temp_c,y=fit,color=P),size=1)+
  xlab("Soil temperature (mean-centred)")+ylab("Predicted fitness")+
  my_theme()+
  facet_wrap(~P,scales="free",labeller=labeller(P=P.labs))
ggsave(filename="output/figures/seeds_temp_c_P_Aster.tiff",device="tiff",
       width=16,height=8,units="cm",dpi=300,compression="lzw")
```

For non-pollinated plants, fitness is higher at cold soil temperatures, but for pollinated plants, fitness is higher at warm soil temperatures.

```{r}
aster_predict1.df.n_seed%>%
  filter(temp_c==min(temp_c)|temp_c==max(temp_c))%>%
  dplyr::select(temp_c,P,fit)
```

#### seeds:temp_c:FFD_c

```{r}
# make prediction df
aster_predict2 <- data.frame(
  expand_grid(temp_c = seq(from = -5, to = 13, length.out = 100),
              FFD_c = seq(from = -29, to = 21, length.out = 100)),
  #FFD_c = 0, # centred so mean=0
  ros_area_c = 0, # centred so mean=0
  P = as.factor(0), # no pollination
  F = as.factor(0),
  seedpods_01 = 1,
  n_seedpods = 1,
  n_seeds = 0,
  root = 1
)

# reshape data to long format
aster_predict2_long <- reshape(as.data.frame(aster_predict2), 
                              varying = list(vars), direction = "long", 
                              timevar = "varb", times = as.factor(vars), 
                              v.names = "resp")

# add artifice
aster_predict2_long$seeds <- as.integer(ifelse(
  aster_predict2_long$varb == "n_seeds", 1, 0))

aster_predict2.df <- aster_predict2_long

# fixed model predictions
aster3.p2 <- predict(aster3, newdata = aster_predict2_long,
                    varvar = varb,idvar = id, root = root,
                    info.tol = 1e-11, se.fit = T)

aster_predict2.df$fit <- aster3.p2$fit
aster_predict2.df$se <- aster3.p2$se.fit

aster_predict2.df.n_seed <- aster_predict2.df %>%
  filter(varb == "n_seeds")

aster_predict2.df.n_seed<-aster_predict2.df.n_seed%>%
  mutate(ymin=fit-se,ymax=fit+se)

# Fitness in normal scale
ggplot(aster_predict2.df.n_seed,aes(x=FFD_c,y=fit,group=as.factor(temp_c)))+
  geom_line(aes(x=FFD_c,y=fit,color=temp_c),size=0.3)+
  xlab("FFD (mean-centred)")+ylab("Predicted fitness")+
  my_theme_legend()+labs(color="Soil temperature\n(mean-centred, ºC)")+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))
ggsave(filename="output/figures/seeds_temp_c_FFD_c_Aster_1.tiff",device="tiff",
       width=15,height=8,units="cm",dpi=300,compression="lzw")

# Fitness in log10 scale
ggplot(aster_predict2.df.n_seed,aes(x=FFD_c,y=fit,group=as.factor(temp_c)))+
  geom_line(aes(x=FFD_c,y=fit,color=temp_c),size=0.3)+
  xlab("FFD (mean-centred)")+ylab("Predicted fitness")+
  my_theme_legend()+scale_y_log10()+
  labs(color="Soil temperature\n(mean-centred, ºC)")+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))
ggsave(filename="output/figures/seeds_temp_c_FFD_c_Aster_1_log.tiff",device="tiff",
       width=15,height=8,units="cm",dpi=300,compression="lzw")

# Fitness in normal scale
ggplot(aster_predict2.df.n_seed,aes(x=temp_c,y=fit,group=as.factor(FFD_c)))+
  geom_line(aes(x=temp_c,y=fit,color=FFD_c),size=0.3)+
  xlab("Soil temperature (mean-centred)")+ylab("Predicted fitness")+
  my_theme_legend()+labs(color="FFD\n(mean-centred)")+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))
ggsave(filename="output/figures/seeds_temp_c_FFD_c_Aster_2.tiff",device="tiff",
       width=15,height=8,units="cm",dpi=300,compression="lzw")

# Fitness in log10 scale
ggplot(aster_predict2.df.n_seed,aes(x=temp_c,y=fit,group=as.factor(FFD_c)))+
  geom_line(aes(x=temp_c,y=fit,color=FFD_c),size=0.3)+
  xlab("Soil temperature (mean-centred)")+ylab("Predicted fitness")+
  my_theme_legend()+scale_y_log10()+labs(color="FFD\n(mean-centred)")+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))
ggsave(filename="output/figures/seeds_temp_c_FFD_c_Aster_2_log.tiff",device="tiff",
       width=15,height=8,units="cm",dpi=300,compression="lzw")
```

It is much better to flower early at cold soil temperatures, but not at warm soil temperatures, where it is slightly better to flower late. 

```{r}
aster_predict2.df.n_seed%>%
  dplyr::select(temp_c,FFD_c,fit)%>%
  filter(temp_c>3.2&temp_c<3.8)
# selection favours earlier flowering at temp_c up to 3.4, 
# and later flowering at higher temp_c
3.4+mean(redata$temp)
# temp_c=3.4 corresponds to temp=16.9
ggplot(subset(aster_predict2.df.n_seed,temp_c>3.2&temp_c<3.8),
       aes(x=FFD_c,y=fit,group=as.factor(temp_c)))+
  geom_line(aes(x=FFD_c,y=fit,color=temp_c),size=0.3)+
  xlab("FFD (mean-centred)")+ylab("Predicted fitness")+
  my_theme_legend()+labs(color="Soil temperature\n(mean-centred, ºC)")+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))
```

#### HERE: seeds:temp_c:FFD_c:P 

```{r}
# # make prediction df
# aster_predict3 <- data.frame(
#   expand_grid(temp_c = seq(from = -5, to = 13, length.out = 100),
#               FFD_c = seq(from = -29, to = 21, length.out = 100),
#               P = as.factor(c(0,1))),
#   ros_area_c = 0, # centred so mean=0
#   F = as.factor(0),
#   seedpods_01 = 1,
#   n_seedpods = 1,
#   n_seeds = 0,
#   root = 1
# )
# 
# # reshape data to long format
# aster_predict3_long <- reshape(as.data.frame(aster_predict3), 
#                               varying = list(vars), direction = "long", 
#                               timevar = "varb", times = as.factor(vars), 
#                               v.names = "resp")
# 
# # add artifice
# aster_predict3_long$seeds <- as.integer(ifelse(
#   aster_predict3_long$varb == "n_seeds", 1, 0))
# 
# aster_predict3.df <- aster_predict1_long
# 
# # fixed model predictions
# aster3.p3 <- predict(aster3, newdata = aster_predict3_long,
#                     varvar = varb,idvar = id, root = root,
#                     parm.type = "canonical", # Works but meaning?
#                     info.tol = 1e-11, se.fit = T)
# 
# aster_predict3.df$fit <- aster3.p3$fit
# aster_predict3.df$se <- aster3.p3$se.fit
# 
# aster_predict3.df.n_seed <- aster_predict3.df %>%
#   filter(varb == "n_seeds")
# 
# aster_predict3.df.n_seed<-aster_predict3.df.n_seed%>%
#   mutate(ymin=fit-se,ymax=fit+se)
# 
# # P.labs <- c("Non-pollinated","Pollinated")
# # names(P.labs) <- c(0,1)
# # ggplot(aster_predict1.df.n_seed)+
# #   geom_ribbon(aes(x=temp_c,ymin=ymin,ymax=ymax,fill=P),alpha=0.25)+
# #   geom_line(aes(x=temp_c,y=fit,color=P),size=1)+
# #   xlab("Soil temperature (mean-centred)")+ylab("Predicted fitness")+
# #   my_theme()+
# #   facet_wrap(~P,scales="free",labeller=labeller(P=P.labs))
# # ggsave(filename="output/figures/seeds_temp_c_P_Aster.tiff",device="tiff",
# #        width=16,height=8,units="cm",dpi=300,compression="lzw")
```

## Figures 2-3

### Figure 2

#### Without points

```{r fig.height=6, fig.width=6}
fig2a<-ggplot(predict_relfitness3,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
    geom_ribbon(alpha=0.2)+
    geom_line(size=1)+
    xlab("Soil temperature (ºC)")+ylab("Predicted relative fitness")+
    my_theme()+ggtitle("A)")+
    facet_wrap(~group,scales="free",labeller=labeller(group=P.labs))
fig2b<-ggplot(aster_predict1.df.n_seed)+
    geom_ribbon(aes(x=temp_c,ymin=ymin,ymax=ymax),alpha=0.25)+
    geom_line(aes(x=temp_c,y=fit),size=1)+
    xlab("Soil temperature (mean-centred, ºC)")+ylab("Predicted fitness")+
    my_theme()+ggtitle("B)")+
    facet_wrap(~P,scales="free",labeller=labeller(P=P.labs))
(fig2<-plot_grid(fig2a, fig2b, ncol=1,align = "v"))
ggsave(fig2,filename="output/figures/fig2.tiff",device="tiff",
       width=16,height=17,units="cm",dpi=300,compression="lzw")
```

#### With points

```{r}
fig2apoints<-
  plot_grid(
  ggdraw()+draw_label("Predicted relative fitness",y = 0.5,angle=90,
                      fontfamily="serif",size=18),
  plot_grid(
    plot_grid(
      ggplot(subset(ping_20_data,P==0),aes(x=temp,y=n_seeds_rel))+
        geom_point(size=2,alpha=0.2)+
        geom_ribbon(data=subset(predict_relfitness3,group==0),
                    aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                    alpha=0.2)+
        geom_line(data=subset(predict_relfitness3,group==0),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),size=1)+
        xlab(NULL)+ylab(NULL)+
        my_theme()+ggtitle("Non-supplemented")+
        theme(plot.title = element_text(hjust = 0.5)),
      ggplot(subset(ping_20_data,P==1),aes(x=temp,y=n_seeds_rel))+
        geom_point(size=2,alpha=0.2)+
        geom_ribbon(data=subset(predict_relfitness3,group==1),
                    aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
                    alpha=0.2)+
        geom_line(data=subset(predict_relfitness3,group==1),
                  aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),size=1)+
        xlab(NULL)+ylab(NULL)+
        my_theme()+ggtitle("Supplemented")+
        theme(plot.title = element_text(hjust = 0.5)),
      ncol=2),
    ggdraw()+draw_label("Soil temperature (ºC)",x = 0.5,
                        fontfamily="serif",size=18),
    ncol=1,rel_heights=c(1.5, 0.1)),
  ncol=2,rel_widths=c(0.1,1.5)
)
fig2bpoints<-
  plot_grid(
  ggdraw()+draw_label("Predicted fitness",y = 0.5,angle=90,
                      fontfamily="serif",size=18),
  plot_grid(
    plot_grid(
        ggplot()+geom_point(data=subset(redata,varb=="n_seeds"&P==0),
                            aes(x=temp_c,y=resp),size=2,alpha=0.2)+
          geom_ribbon(data=subset(aster_predict1.df.n_seed,P==0),
                      aes(x=temp_c,ymin=ymin,ymax=ymax),alpha=0.25)+
          geom_line(data=subset(aster_predict1.df.n_seed,P==0),
                    aes(x=temp_c,y=fit),size=1)+
          xlab(NULL)+ylab(NULL)+
          my_theme()+ggtitle("Non-supplemented")+
          theme(plot.title = element_text(hjust = 0.5)),
        ggplot()+geom_point(data=subset(redata,varb=="n_seeds"&P==1),
                            aes(x=temp_c,y=resp),size=2,alpha=0.2)+
          geom_ribbon(data=subset(aster_predict1.df.n_seed,P==1),
                      aes(x=temp_c,ymin=ymin,ymax=ymax),alpha=0.25)+
          geom_line(data=subset(aster_predict1.df.n_seed,P==1),
                    aes(x=temp_c,y=fit),size=1)+
          xlab(NULL)+ylab(NULL)+
          my_theme()+ggtitle("Supplemented")+
          theme(plot.title = element_text(hjust = 0.5)),
        ncol=2),
    ggdraw()+draw_label("Soil temperature (mean-centred, ºC)",x = 0.5,
                        fontfamily="serif",size=18),
    ncol=1,rel_heights=c(1.5, 0.1)),
  ncol=2,rel_widths=c(0.1,1.5)
)
(fig2points<-plot_grid(fig2apoints, fig2bpoints, ncol=1,align = "v"))
ggsave(fig2points,filename="output/figures/fig2points.tiff",device="tiff",
       width=16,height=17,units="cm",dpi=300,compression="lzw")
```


### Figure 3

```{r fig.height=6, fig.width=12}
fig3a<-ggplot(predict_relfitness1,aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high,
                            group=group))+
  geom_line(aes(color=as.numeric(as.character(group))),size=0.3)+
  xlab("FFD (standardized)")+ylab("Predicted relative fitness")+
  my_theme_legend()+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))+
  labs(colour="Soil  temperature (ºC)")+
  theme(legend.position="top")+ggtitle("A)")
fig3b<-ggplot(aster_predict2.df.n_seed,aes(x=FFD_c,y=fit,group=as.factor(temp_c)))+
  geom_line(aes(x=FFD_c,y=fit,color=temp_c),size=0.3)+
  xlab("FFD (mean-centred)")+ylab("Predicted fitness")+
  my_theme_legend()+scale_y_log10()+
  labs(color="Soil temperature\n(mean-centred, ºC)")+
  scale_colour_gradientn(colours = plasma(n=100,begin=0.4,end=0.9,direction=-1))+
  theme(legend.position="top")+ggtitle("B)")
(fig3<-plot_grid(fig3a, fig3b,align="h",rel_widths=c(1, 1.08)))
ggsave(fig3,filename="output/figures/fig3.tiff",device="tiff",
       width=19,height=11,units="cm",dpi=300,compression="lzw")
```

# A. Analyses 2020-2021

## 1. Models for FFD 

Effect of temperature on FFD

Using linear models, including quadratic effects of temp.

```{r echo=TRUE}
FFD_2021_1<-lm(FFD_corr~temp+I(temp^2),ping_21_data)
summary(FFD_2021_1)
```

Model diagnostics (not shown) indicated that linear models are OK.

### Model prediction

```{r, fig.height=3, fig.width=4}
predict_FFD_21<-ggpredict(FFD_2021_1,terms="temp")
ggplot(ping_21_data,aes(x=temp,y=FFD_corr))+
  geom_ribbon(data=predict_FFD_21,
              aes(x=x,y=predicted,ymin=conf.low,ymax=conf.high),
              fill="grey",alpha=0.7)+
  geom_line(data=predict_FFD_21,aes(x=x,y=predicted),size=1,color="black")+
  geom_point(size=2,alpha=0.2)+
  xlab("Soil temperature (ºC)")+ylab("FFD")+
  my_theme()
```

## 2. Carry-over effects

Merge data 2020 with useful data 2021

```{r}
ping_21_data_useful<-ping_21_data%>%
  dplyr::select(id,FFD_corr,temp,n_stems,n_opened_fl,n_seedpods,
                max_h,med_h,flowered,ros_area,fitness_01)%>%
  rename(FFD_corr_21=FFD_corr,temp_21=temp,n_stems_21=n_stems,
         n_opened_fl_21=n_opened_fl,n_seedpods_21=n_seedpods,
         max_h_21=max_h,med_h_21=med_h,flowered_21=flowered,
         ros_area_21=ros_area,fitness_01_21=fitness_01)
ping_20_21_data<-left_join(ping_20_data,ping_21_data_useful)
```

```{r}
nrow(subset(ping_20_21_data,!is.na(FFD_corr_21)))
nrow(subset(ping_20_21_data,!is.na(flowered_21)))
nrow(subset(ping_20_21_data,flowered_21==1))
```
140 plants with treatment in 2020 that flowered in 2021.

```{r}
summary(lm(FFD_corr_21~P,ping_20_21_data))
summary(lm(FFD_corr_21~F,ping_20_21_data))
summary(glm(n_stems_21~P,ping_20_21_data,family="poisson"))
summary(glm(n_stems_21~F,ping_20_21_data,family="poisson"))
summary(glm(n_opened_fl_21~P,ping_20_21_data,family="poisson"))
summary(glm(n_opened_fl_21~F,ping_20_21_data,family="poisson"))
summary(glm(n_seedpods_21~P,ping_20_21_data,family="poisson"))
summary(glm(n_seedpods_21~F,ping_20_21_data,family="poisson"))
summary(lm(max_h_21~P,ping_20_21_data))
summary(lm(max_h_21~F,ping_20_21_data)) #* Opposite!
summary(lm(med_h_21~P,ping_20_21_data))
summary(lm(med_h_21~F,ping_20_21_data)) #* Opposite!
summary(glm(flowered_21~P,ping_20_21_data,family="binomial"))
summary(glm(flowered_21~F,ping_20_21_data,family="binomial"))
summary(lm(log(ros_area_21)~P,ping_20_21_data)) #*
summary(lm(log(ros_area_21)~F,ping_20_21_data))
summary(glm(fitness_01_21~P,ping_20_21_data,family="binomial")) 
summary(glm(fitness_01_21~F,ping_20_21_data,family="binomial"))

summary(lm(FFD_corr_21~treatment,ping_20_21_data))
summary(lm(n_stems_21~treatment,ping_20_21_data))
summary(lm(n_opened_fl_21~treatment,ping_20_21_data))
summary(lm(n_seedpods_21~treatment,ping_20_21_data))
summary(lm(max_h_21~treatment,ping_20_21_data)) #* Opposite!
summary(lm(med_h_21~treatment,ping_20_21_data)) #* Opposite!
summary(glm(flowered_21~treatment,ping_20_21_data,family="binomial"))
summary(lm(log(ros_area_21)~treatment,ping_20_21_data))
summary(glm(fitness_01_21~treatment,ping_20_21_data,family="binomial"))

ggplot(ping_20_21_data,aes(F,max_h_21))+geom_boxplot()
ggplot(ping_20_21_data,aes(treatment,max_h_21))+geom_boxplot()
ggplot(ping_20_21_data,aes(F,med_h_21))+geom_boxplot()
ggplot(ping_20_21_data,aes(treatment,med_h_21))+geom_boxplot()
ggplot(ping_20_21_data,aes(P,ros_area_21))+geom_boxplot()
ggplot(ping_20_21_data,aes(treatment,ros_area_21))+geom_boxplot()

summary(lm(max_h~F,ping_20_21_data)) #NS but similar trend
summary(lm(med_h~F,ping_20_21_data)) #NS but similar trend

with(ping_20_21_data,cor(max_h,max_h_21,use="complete.obs")) # moderate
with(ping_20_21_data,cor(med_h,med_h_21,use="complete.obs")) # moderate

with(ping_20_21_data,cor(temp,temp_21,use="complete.obs")) # high
```

```{r}
ping_20_21_data<-ping_20_21_data%>%
  mutate(FFD_corr_21_std=as.numeric(scale(FFD_corr_21)),
         ros_area_21_std=as.numeric(scale(log(ros_area_21))))

subset4<-subset(ping_20_21_data,!is.na(fitness_01_21)&!is.na(ros_area_21_std))
globmod_selection_2020_21_01<-glm(fitness_01_21~temp_21*FFD_corr_21_std*P*F+
                                    ros_area_21_std,subset4,
                                  family="binomial",na.action="na.fail")
clusterType <- if(length(find.package("snow", quiet = TRUE))) "SOCK" else "PSOCK"
clust1 <- try(makeCluster(getOption("cl.cores", 3), type = clusterType))
clusterExport(clust1, "subset4")
clusterEvalQ(clust1, library(stats))

modsel_selection_2020_21_01<-pdredge(globmod_selection_2020_21_01,
                                 fixed=c("ros_area_21_std"),
                                 cluster=clust1)

summary(model.avg(modsel_selection_2020_21_01,
                  subset=delta<2)) # Summary averaged model
importance(modsel_selection_2020_21_01) # Variable importance
r.squaredGLMM(get.models(modsel_selection_2020_21_01,subset=1)$"2") 
#R square of best model
```

```{r}
summary(glm(fitness_01_21~temp_21*FFD_corr_21_std+ros_area_21_std,
            ping_20_21_data,family="binomial"))
```


# Session info

```{r eval=FALSE, include=FALSE}
sessionInfo()
```
